requirements.txt
cat > requirements.txt << EOF
kivy==2.3.0
pillow==10.0.0
EOF